style: add utils for style shorthands
